+++ lib/faker.dart
@@ -1,37 +1,34 @@
 /// A library for Dart that generates fake data.
-library faker;
 
-import 'dart:math';
+export 'src/faker.dart';
 
-part 'src/faker.dart';
+export 'src/random_generator.dart';
 
-part 'src/random_generator.dart';
+export 'src/address.dart';
+export 'src/conference.dart';
+export 'src/company.dart';
+export 'src/currency.dart';
+export 'src/food.dart';
+export 'src/guid.dart';
+export 'src/internet.dart';
+export 'src/job.dart';
+export 'src/person.dart';
+export 'src/sport.dart';
 
-part 'src/address.dart';
-part 'src/conference.dart';
-part 'src/company.dart';
-part 'src/currency.dart';
-part 'src/food.dart';
-part 'src/guid.dart';
-part 'src/internet.dart';
-part 'src/job.dart';
-part 'src/person.dart';
-part 'src/sport.dart';
-
-part 'src/data/address/city_suffixes.dart';
-part 'src/data/address/countries.dart';
-part 'src/data/address/country_codes.dart';
-part 'src/data/address/neighborhoods.dart';
-part 'src/data/address/street_suffixes.dart';
-part 'src/data/conference/conference_names.dart';
-part 'src/data/currency/currency_names.dart';
-part 'src/data/currency/currency_codes.dart';
-part 'src/data/food/cuisines.dart';
-part 'src/data/food/dishes.dart';
-part 'src/data/food/restaurants.dart';
-part 'src/data/person/firstnames.dart';
-part 'src/data/person/lastnames.dart';
-part 'src/data/job/job_adj.dart';
-part 'src/data/job/job_noun.dart';
-part 'src/data/job/job_prefix.dart';
-part 'src/data/sport/sport_names.dart';
+export 'src/data/address/city_suffixes.dart';
+export 'src/data/address/countries.dart';
+export 'src/data/address/country_codes.dart';
+export 'src/data/address/neighborhoods.dart';
+export 'src/data/address/street_suffixes.dart';
+export 'src/data/conference/conference_names.dart';
+export 'src/data/currency/currency_names.dart';
+export 'src/data/currency/currency_codes.dart';
+export 'src/data/food/cuisines.dart';
+export 'src/data/food/dishes.dart';
+export 'src/data/food/restaurants.dart';
+export 'src/data/person/firstnames.dart';
+export 'src/data/person/lastnames.dart';
+export 'src/data/job/job_adj.dart';
+export 'src/data/job/job_noun.dart';
+export 'src/data/job/job_prefix.dart';
+export 'src/data/sport/sport_names.dart';