+++ lib/src/address.dart
@@ -1,4 +1,11 @@
-part of faker;
+import 'random_generator.dart';
+import 'faker.dart';
+import 'data/address/street_suffixes.dart';
+import 'data/address/city_suffixes.dart';
+import 'data/address/neighborhoods.dart';
+import 'data/address/countries.dart';
+import 'data/address/country_codes.dart';
+import 'data/address/continents.dart';
 
 class Address {
   static const compassDirections = const ['North', 'East', 'West', 'South'];
@@ -42,9 +49,9 @@ class Address {
   ///   faker.address.cityPrefix();
   /// ```
   String cityPrefix() => [
-    random.element(compassDirections),
-    random.element(cityPrefixes)
-  ].join(' ');
+        random.element(compassDirections),
+        random.element(cityPrefixes)
+      ].join(' ');
 
   /// Generates a city suffix.
   ///
@@ -111,4 +118,12 @@ class Address {
   ///   faker.address.countryCode();
   /// ```
   String countryCode() => random.element(countryCodes);
+
+  /// Generates a continent.
+  ///
+  /// Example:
+  /// ```dart
+  ///   faker.address.continent();
+  /// ```
+  String continent() => random.element(continents);
 }