+++ lib/src/conference.dart
@@ -1,4 +1,5 @@
-part of faker;
+import 'random_generator.dart';
+import 'data/conference/conference_names.dart';
 
 class Conference {
   const Conference();