+++ lib/src/data/address/countries.dart
@@ -1,5 +1,3 @@
-part of faker;
-
 const countries = const [
   "Afghanistan",
   "Åland Islands",