+++ lib/src/data/address/street_suffixes.dart
@@ -1,5 +1,3 @@
-part of faker;
-
 const streetSuffixes = const [
   'Alley',
   'Avenue',