+++ lib/src/faker.dart
@@ -1,4 +1,16 @@
-part of faker;
+import 'package:faker/src/lorem.dart';
+
+import 'address.dart';
+import 'company.dart';
+import 'conference.dart';
+import 'currency.dart';
+import 'food.dart';
+import 'guid.dart';
+import 'internet.dart';
+import 'job.dart';
+import 'person.dart';
+import 'random_generator.dart';
+import 'sport.dart';
 
 const Faker faker = const Faker();
 
@@ -11,6 +23,7 @@ class Faker {
   final Guid guid;
   final Internet internet;
   final Job job;
+  final Lorem lorem;
   final Person person;
   final Sport sport;
   final RandomGenerator randomGenerator;
@@ -24,6 +37,7 @@ class Faker {
         guid = const Guid(),
         internet = const Internet(),
         job = const Job(),
+        lorem = const Lorem(),
         person = const Person(),
         sport = const Sport(),
         randomGenerator = const RandomGenerator();