+++ lib/src/guid.dart
@@ -1,10 +1,8 @@
-part of faker;
+import 'random_generator.dart';
 
 class Guid {
-
   const Guid();
 
-  String guid() => random.fromPatternToHex(
-      ['########-####-####-####-############']
-  );
+  String guid() =>
+      random.fromPatternToHex(['########-####-####-####-############']);
 }