+++ test/runner.dart
@@ -6,6 +6,7 @@ import 'specs/food.dart' as food;
 import 'specs/guid.dart' as guid;
 import 'specs/internet.dart' as internet;
 import 'specs/job.dart' as job;
+import 'specs/lorem.dart' as lorem;
 import 'specs/person.dart' as person;
 import 'specs/sport.dart' as sport;
 import 'specs/random_generator.dart' as random;
@@ -19,6 +20,7 @@ main() {
   guid.main();
   internet.main();
   job.main();
+  lorem.main();
   person.main();
   sport.main();
   random.main();