+++ test/specs/internet.dart
@@ -1,79 +1,64 @@
-library faker.test.internet;
-
-import 'package:guinness/guinness.dart';
+import 'package:test/test.dart';
 import 'package:faker/faker.dart';
 
 main() {
   var faker = new Faker();
 
-  describe('internet', () {
-    it('should be able to generate email', () {
-      expect(new RegExp(r'[a-zA-Z.+]+@[a-zA-Z.]+')
-          .hasMatch(faker.internet.email())).toBeTrue();
+  group('internet', () {
+    test('should be able to generate email', () {
+      expect(faker.internet.email(), matches(new RegExp(r'[a-zA-Z.+]+@[a-zA-Z.]+')));
     });
 
-    it('should be able to generate disposable email', () {
-      expect(new RegExp(r'[a-zA-Z.+]+@[a-zA-Z.]+')
-          .hasMatch(faker.internet.disposableEmail())).toBeTrue();
+    test('should be able to generate disposable email', () {
+      expect(faker.internet.disposableEmail(), matches(new RegExp(r'[a-zA-Z.+]+@[a-zA-Z.]+')));
     });
 
-    it('should be able to generate free email', () {
-      expect(new RegExp(r'[a-zA-Z.+]+@[a-zA-Z.]+')
-          .hasMatch(faker.internet.freeEmail())).toBeTrue();
+    test('should be able to generate free email', () {
+      expect(faker.internet.freeEmail(), matches(new RegExp(r'[a-zA-Z.+]+@[a-zA-Z.]+')));
     });
 
-    it('should be able to generate safe email', () {
-      expect(new RegExp(r'[a-zA-Z.+]+@[a-zA-Z.]+')
-          .hasMatch(faker.internet.safeEmail())).toBeTrue();
+    test('should be able to generate safe email', () {
+      expect(faker.internet.safeEmail(), matches(new RegExp(r'[a-zA-Z.+]+@[a-zA-Z.]+')));
     });
 
-    it('should be able to generate user name', () {
-      expect(new RegExp(r'[a-z]+((_|-|\.)[a-z]+)')
-          .hasMatch(faker.internet.userName())).toBeTrue();
+    test('should be able to generate user name', () {
+      expect(faker.internet.userName(), matches(new RegExp(r'[a-z]+((_|-|\.)[a-z]+)')));
     });
 
-    it('should be able to generate domain name', () {
-      expect(new RegExp(r'^((?!-)[A-Za-z0-9-]{1,63}).+[A-Za-z]{2,6}$')
-          .hasMatch(faker.internet.domainName())).toBeTrue();
+    test('should be able to generate domain name', () {
+      expect(faker.internet.domainName(), matches(new RegExp(r'^((?!-)[A-Za-z0-9-]{1,63}).+[A-Za-z]{2,6}$')));
     });
 
-    it('should be able to generate domain word', () {
-      expect(new RegExp(r"^[\w']+$").hasMatch(faker.internet.domainWord()))
-          .toBeTrue();
+    test('should be able to generate domain word', () {
+      expect(faker.internet.domainWord(), matches(new RegExp(r"^[\w']+$")));
     });
 
-    it('should be able to generate uri', () {
-      expect(new RegExp(r'^(ftp://.+)$').hasMatch(faker.internet.uri('ftp')))
-          .toBeTrue();
+    test('should be able to generate uri', () {
+      expect(faker.internet.uri('ftp'), matches(new RegExp(r'^(ftp://.+)$')));
     });
 
-    it('should be able to generate http url', () {
-      expect(new RegExp(r'^(http://.+)$').hasMatch(faker.internet.httpUrl()))
-          .toBeTrue();
+    test('should be able to generate http url', () {
+      expect(faker.internet.httpUrl(), matches(new RegExp(r'^(http://.+)$')));
     });
 
-    it('should be able to generate https url', () {
-      expect(new RegExp(r'^(https://.+)$').hasMatch(faker.internet.httpsUrl()))
-          .toBeTrue();
+    test('should be able to generate https url', () {
+      expect(faker.internet.httpsUrl(), matches(new RegExp(r'^(https://.+)$')));
     });
 
-    it('should be able to generate ipv4 address', () {
-      expect(Uri.parseIPv4Address(faker.internet.ipv4Address()).length)
-          .toEqual(4);
+    test('should be able to generate ipv4 address', () {
+      expect(Uri.parseIPv4Address(faker.internet.ipv4Address()), hasLength(4));
     });
 
-    it('should be able to generate ipv6 address', () {
-      expect(Uri.parseIPv6Address(faker.internet.ipv6Address()).length)
-          .toEqual(16);
+    test('should be able to generate ipv6 address', () {
+      expect(Uri.parseIPv6Address(faker.internet.ipv6Address()), hasLength(16));
     });
 
-    it('should be able to generate mac address', () {
-      expect(new RegExp(r'^([\da-f]{2}[:]){5}([\da-f]{2})$')
-          .hasMatch(faker.internet.macAddress())).toBeTrue();
+    test('should be able to generate mac address', () {
+      expect(faker.internet.macAddress(), matches(new RegExp(r'^([\da-f]{2}[:]){5}([\da-f]{2})$')));
     });
 
-    it('should be able to generate password', () {
-      expect(faker.internet.password(length: 40).length).toEqual(40);
+    test('should be able to generate password', () {
+      expect(faker.internet.password(length: 40), hasLength(40));
     });
   });
 }